refactor: add channel-card module. channel-card and card-tvshow components make as shared
