+++ src/app/shared/components/channel-card/channel-card.component.scss
