+++ src/app/shared/components/channel-card/components/card-tvshow/card-tvshow.component.scss
