+++ src/app/shared/components/channel-card/components/card-tvshow/card-tvshow.component.ts
@@ -9,8 +9,8 @@ import {
 import { BehaviorSubject } from 'rxjs';
 import * as moment from 'moment';
 import { Moment } from 'moment';
-import { TvshowModel } from '../../../../../../shared/interfaces/tvshow.model';
-import { ProcessingService } from '../../../../../../shared/services/processing.service';
+import { TvshowModel } from '../../../../interfaces/tvshow.model';
+import { ProcessingService } from '../../../../services/processing.service';
 
 @Component({
   selector: 'app-card-tvshow',
